refractor: api -> apiUtility
